chore: release version 0.74.0-alpha.1
The main branch has diverged significantly from the public testnet, so
I'm bumping the version on main to denote that its code represents an
as-yet-unreleased version, 0.74.0. There's no tag here, so it won't
trigger a release workflow. It's just to update log messages and make it
easier to reason about versions across the SDLC.