dex: fix frontier cache incoherence (#4283)
**Description**

This PR fix a bug in the `Frontier` position cache that caused a
position execution guard to be hit.

**Context**

During routing execution, the DEX engine assemble frontiers of positions
to execute against. It maintains an internal cache with a position
stream for each hop in the path, along with the ids and state of the
current loaded set. The position cache is persisted in its entirety when
a round of routing execution has completed (via `Frontier::save`), and
partially, if a position needs to be replaced
(`Frontier::replace_position`).

**Bug**

The bug occurs when the following conditions align:
- a position needs to be replaced
- that position is auto-closing, and set to `Closed` when recording its
execution
- no replacements are available

In that case, since the position rotation logic does not update the
internal state of the position, it will attempt to record execution of a
stale position when saving the frontier.

**Fix**

The fix is two-fold, first, `PositionManager::position_execution` should
make it clear that it can mutate its input and return the updated
position state to allow callers to inspect its state. Once that's done,
the position rotation logic will be able to immediately update its
internal cache with the actual position state so that there are no
incoherences between its internal view and the actual chain state.


## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > This is consensus breaking.