auction: add views to the `ActionView` definition (#4295)
## Describe your changes

This PR adds `ActionDutchAuctionScheduleView` and
`ActionDutchAuctionWithdrawView` to the `ActionView` message, it also
replaces the `IsAction` perspective impls with concrete implementation.

Populating a view for the withdraw action without extending the `TxP`
structure has proven squirrelly, and it will require a follow-up,
although the data is "here" in the view server so I don't think it will
be too complicated.

## Issue ticket number and link

Part of #4219 

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> Breaking changes are made to `ActionView`: the
`action_dutch_auction_schedule` and `action_dutch_auction_withdraw`
fields have had their types changes from `ActionDutchAuctionSchedule`
and `ActionDutchAuctionWithdraw` to `ActionDutchAuctionScheduleView` and
`ActionDutchAuctionWithdrawView`, respectively.