view: port wasm planner to core  (#4312)
## Describe your changes

This PR adds a port of the wasm planner, integrating it within the
`penumbra-view` crate.

**`SwapClaim` handling**
In the process of porting this PR, we found a bug in the handling of
`SwapClaim` actions with the wasm planner.

`SwapClaim` actions produce surplus value from their prepaid fees) that
must be released so that the transaction balance to zero. In principle,
it should be possible to release this value into an output note. This
isn't the approach we took with this PR. Instead, we release it to the
transaction fee, providing a base for us to improve on later. This will
be tracked by https://github.com/penumbra-zone/penumbra/issues/4313.

## Issue ticket number and link
References https://github.com/penumbra-zone/penumbra/issues/4081

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

---------

Co-authored-by: Erwan Or <erwanor@penumbralabs.xyz>