view: simplify delegator vote logic in Planner
The existing logic didn't balance delegator vote transactions properly.  This
commit is the first step to fixing it.  The purpose of allowing the `Planner`
to have an async API was so that it could fetch data directly from the view
service, rather than having to implement our own async logic at the application
level (by saving stuff we want to query later).

However, the new planner logic doesn't seem to be correct, because adding the
`DelegatorVotePlan` gives an error that there were no actions to balance.