view: restructure `Planner` to use an `ActionList`
The part of the web planner refactoring we wanted to bring over to the core
repo was the creation of a separate `ActionList` modeling a list of actions
that may or may not be balanced.  This commit pulls the planner into that
shape, and extends the `ActionList` design found in the web code in a few ways:

- The `ActionList` tracks a target fee internally, rather than leaving that to
  its caller. This allows us to handle cases like a swap claim, where we want
  to put the prepaid fee released by a swap claim towards the transaction fee,
  and not create dust outputs.

- The `ActionList` has the groundwork for supporting multi-asset fee payments,
  by not assuming that the fee is always in the staking token when accounting
  for fees. Later we can add an asset ID to the `GasPrices` and it should be
  a relatively drop-in extension.

- The `ActionList` can be used to create a `TransactionPlan` directly. In the
  future, we should consider restructuring the `view` crate to have most of
  its current contents behind a `svc` flag, allowing us to depend on the `view`
  crate from the wasm code and have a common `ActionList` impl between them,
  just different code that manipulates it.

Various API cleanups were necessary along the way.