feat(pd): tonic-metrics create PARTIAL
The goal: add metrics for gRPC service methods, to capture at least:

  * number of times a given service was called (gauge)
  * duration of each request (histogram)

The first pass on the shape of this implementation comes from a pairing
session with Henry (and Kate).

Factors to consider:

  * we don't want to get dos'd, so only valid methods should be emitted

Adds a fn to inspect the penumbra-proto crate's FILE_DESCRIPTOR_SET,
so we can validate request routes as valid messages.