view: 👀 introduce more tracing instrumentation (#4349)
## describe your changes

this introduces some additional `tracing` instrumentation, used while
debugging
integration with the view server in mock consensus tests. because these
events
may be useful more broadly, and aren't strictly tied to testing, this is
cherry-picked out of that work for separate consideration.

## issue ticket number and link

* #3913

## checklist before requesting a review

- [x] if this code contains consensus-breaking changes, i have added the
"consensus-breaking" label. otherwise, i declare my belief that there
are not consensus-breaking changes, for the following reason:

this branch only makes changes to tracing spans and events. as this is
only altering telemetry, this is not consensus-breaking.

---------

Co-authored-by: Henry de Valence <hdevalence@penumbralabs.xyz>