proto: DelegatorVoteView has opaque/visible variants  (#4291)
## Describe your changes

Small change to `transaction.proto` that moves `DelegatorVoteView` to be
included in the section of `ActionView`s that are explicitly documented
as having `Opaque` and `Visible` variants. Beyond generated protobuf
code from `scripts/protobuf-codegen`, there are no other changes.

## Issue ticket number and link

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This PR only changes the order of a Protobuf Message definition for
the purposes of documentation.