ci: automate chain upgrades on k8s deployments
Here's long-needed automation for deploying chain upgrades on Penumbra
testnets deployed to k8s. Ideally we'd use an operator pattern to manage
the lifecycle, but we're trying to move quickly, so bash it'll be for
now.

Also included are some lower thresholds for the devnet environment,
since it's frequently used for this kind of testing. Might make more
sense to create a dedicated "upgrade-testing" environment, but again,
prioritizing speed and reliability, so we can circle back on fleshing
things out once we hit our deadlines.

Tacked on some help scripts, such as "enable-maintenance-mode"
and "perform-point-release", which are helpful for managing devnets.
debugging devnets.