ci: fix rust-docs build
Bumped the `time` crate in order to resolve a build error with
recent rust nightlies on the `rust-docs` build. Also removed
the `index.html` file generated for rust-docs from version control,
since that file is built on the fly in CI post-merge, and so will always
be out of date.