Include vendored proto definitions in the proto_descriptor file (#4425)
## Describe your changes

This changes the proto build process to include cosmos/ibc-go proto
files in the vendored proto file list used to generate the
proto_descriptor file.

As-implemented, there are some duplicate paths that appear both as
dependencies of the penumbra protos and in ibc-go. The ibc-go protos
will be included first so the penumbra protos take priority. This
shouldn't be an issue in the way we currently use the protos but this
seems important to document for future reference.

## Issue ticket number and link

Closes #4422

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Just changes gRPC reflection returns