view: 📫 do not thrash `error_slot` mutex
`check_worker()` is called by almost every public interface in the view
server, as a preliminary check that the worker task running in the
background has not recorded an error of some kind.

this commit makes a change to `check_worker()`, so that it does not need
to frivolously acquire, release, and subsequently reacquire the mutex in
the event of an error being stored in the slot.

tracing events and documentation are also added to this function.

x-ref: #3913