mock-tendermint-proxy: 🪩 define tendermint proxy mocks
our broader goal is to wire up plumbing so that a test node can handle
the tendermint proxy service client's requests sent by the view server
when planning transactions.

as a step towards that, add a new crate to the workspace. for the time
being, we provide an empty `Stub` proxy. we will build upon this and
sort out connecting it to a running `TestNode<C>` next.