app: 👐 `rpc::router(..)` accepts a `TendermintProxyService`
 #### 💭 describe your changes

instead of accepting a `url::Url` and constructing specifically the
"real" `TendermintProxy`, tweak the signature of the router function so
that it accepts a `T: TendermintProxyService`.

in the future, this will allow mock consensus test cases to run an rpc
endpoint for the view server, without running a Tendermint/CometBFT
process.

 #### 🔖 issue ticket number and link

see #3913.

 #### ✅ checklist before requesting a review

- [x] if this code contains consensus-breaking changes, i have added the
  "consensus-breaking" label. otherwise, i declare my belief that there are not
  consensus-breaking changes, for the following reason:

    > this is a like-for-like refactoring, changing a function
    > signature.