governance: lower emergency vote threshold to 1/3 (#4440)
This changes the ratio in tally.rs emergency_pass() to 1/3 and also
updates any comments/docs referring to the previous 2/3 threshold. Note:
this will also affect `IBCFreeze` and `IBCUnfreeze`, so those have also
gotten easier to accomplish (seems fine for the same reasons, but wasn't
explicitly noted in the ticket). Closes #4389.