dex: implement streaming candlesticks like compact blocks
This also rearranges some of the existing code to split out the component parts from the data modeling parts.

In the future, we could add more metrics or consider writing something generic
for this pattern of "subscribe and pull out of state", though it might not make
sense until we have a third API since this one is already different from the
compact block RPC (where there's a catch up phase). But that could be a good
way to do things like per-RPC connection counters. Or perhaps that should
better not be done "inside" each streaming RPC but as a middleware that could
be on any streaming RPC.