dex: `RoutingParams: From<DexParameters>`
this is a small change. this is a nice property about our data model,
i.e. that an owned copy of the dex parameters may be used to make
routing parameters.

let's make that relationship explicit by putting it into a `From<T>`
impl, and pulling it out of the `StateReadExt` logic.