cnidarium(watch): always check that the channel is open (#4474)
## Describe your changes

We need to bind server resource usage to the lifetime of a connection.
Currently, the way client connection termination is percolated to the
state watcher task is that sending over the channel would fail as soon
as the receive handler is dropped. However, this is insufficient because
we only send over the channel in specific cases. Instead, we should be
checking unconditionally.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Mechanical change