pd: 💤 a sleep worker, tracking scheduler latency
see the module-level docs of `pd::metrics::sleep_worker`. this
introduces a submodule to `pd::metrics`, which observes tokio scheduler
latency and records it for the Prometheus exporter.

the core idea in this worker is that it will invoke
`tokio::time::sleep(..)` to put the task to sleep for one second. upon
being woken up by the scheduler, it will check the *new* wall-clock time
and calculate the actual amount of time it spent waiting. if this took
longer than expected, it increments the
`pd_async_sleep_drift_milliseconds` counter by the observed latency,
measured in milliseconds.

as the module-level docs note, this is a very useful tool to identify
when the runtime is being disrupted by blocking I/O or other expensive
computation.

pd: 🚪 `metrics` module can be `pub`

this will make it slightly nicer to expose facilities for telemetry in
the `pd` binary, without clouding the top-level namespace.