docs: 🪨 add notes on linking to prebuilt rocksdb
 #### 💭 background

NB: this commit is based on #4495.

`librocksdb-sys(build)` is a somewhat infamous source of lengthy build
times when engineers are working in the monorepo. this is a crate whose
build script, clones and builds the rocksdb database from source,
statically linking against the generated `librocksdb.a` file.

this interacts poorly with a Cargo workspace like ours, which contains a
number of different features. because features are additive, running
commands like `cargo build --workspace` and `cargo build --package penumbra-app`
can result in different feature sets being enabled for leaf
dependencies, mandating that the library be rebuilt (_which means
rebuilding librocksdb.a once more_).