flake: 🔗 link against a precompiled rocksdb
this adds a `ROCKSDB_LIB_DIR` environment variable to the nix
development shell.

this environment variable points to the directory that
contains the `librocksdb.a` file, to be statically linked against the
`librocksdb-sys` crate.

a `nixpkgs.rocksdb` build input is added.

 #### 📐 build timing methodology

this shell script was used to generate build timings. in order to see
how this behaved on an incremental build, the script will clear the
cache, build the workspace, and _then_ build a crate that depends on
`librocksdb-sys`.

```sh
 #!/usr/bin/env bash
 #
 # see https://doc.rust-lang.org/cargo/reference/timings.html for more info.
 set -euox pipefail

 # clean out any preëxisting build artifacts.
 cargo clean;

 # build once before we generate our report.
 cargo build --workspace

 # build a particular crate, recording the build timings.
 cargo build --package cnidarium --timings

 # hang on to the report generated.
 cp target/cargo-timings/cargo-timing.html .
```