crypto: enable trial decryption of legacy swap ciphertexts
After the swap encryption is changed to use a fixed-nonce, previous
swap ciphertexts will not be decryptable. In this commit, we enable
the trial decryption of swap ciphertexts using first the new fixed-
nonce method, and then the `LegacySwap` method, where the prefix
bytes of the swap commitment are used as a nonce.