chore: 🪨 improve incremental builds by linking to rocksdb (#4496)
first, the cool part: here are incremental build timings for `cargo
build -p cnidarium` filtering for compilation units that took > 1.5
seconds to compile.

on my machine, this improved incremental build times by 72.92%

**before** :weary: 

![warm-build-timing-without-rocksdb-linking](https://github.com/penumbra-zone/penumbra/assets/57912822/20e60b9b-8f0c-4834-9cde-de492a52c65e)

**after** :heart_eyes: 

![warm-build-timing-with-rocksdb-linking](https://github.com/penumbra-zone/penumbra/assets/57912822/bf1a72b3-ed6c-47ec-9588-84d484aa1c1a)

## :heavy_plus_sign: changes

### docs: 🪨 add notes on linking to prebuilt rocksdb

#### 💭 background

`librocksdb-sys(build)` is a somewhat infamous source of lengthy build
times when engineers are working in the monorepo. this is a crate whose
build script, clones and builds the rocksdb database from source,
statically linking against the generated `librocksdb.a` file.

this interacts poorly with a Cargo workspace like ours, which contains a
number of different features. because features are additive, running
commands like `cargo build --workspace` and `cargo build --package
penumbra-app`
can result in different feature sets being enabled for leaf
dependencies, mandating that the library be rebuilt (_which means
rebuilding librocksdb.a once more_).

### flake: 🔗 link against a precompiled rocksdb

this adds a `ROCKSDB_LIB_DIR` environment variable to the nix
development shell.

this environment variable points to the directory that
contains the `librocksdb.a` file, to be statically linked against the
`librocksdb-sys` crate.

a `nixpkgs.rocksdb` build input is added.

#### 📐 build timing methodology

this shell script was used to generate build timings. in order to see
how this behaved on an incremental build, the script will clear the
cache, build the workspace, and _then_ build a crate that depends on
`librocksdb-sys`.

```sh
 #!/usr/bin/env bash
 #
 # see https://doc.rust-lang.org/cargo/reference/timings.html for more info.
 set -euox pipefail

 # clean out any preëxisting build artifacts.
 cargo clean;

 # build once before we generate our report.
 cargo build --workspace

 # build a particular crate, recording the build timings.
 cargo build --package cnidarium --timings

 # hang on to the report generated.
 cp target/cargo-timings/cargo-timing.html .
```