fee: swap alt fee tokens into the native asset for distribution
This commit moves fee handling out of the Transaction's ActionHandler and into
the component, which accumulates fees (in any token), and adds hooks to the DEX
thatadd chain-submitted swaps for any non-native fee tokens into the native
token. These are then accumulated back into the fee component.

The base fee and tip are tracked separately through this whole process, in
order to support the intended behaviour (the base fee is burned to account for
the _chain_'s resource use, the tip is sent to the proposer to encourage them
to include transactions in fee priority order). However, tip handling is not
currently implemented, so both are burned.  Later, the funding/distribution
components should extract the tip and send it to the proposer's funding
streams. However, until a robust fee market develops, this form of proposer
incentivization can be deferred. The accounting will be there for it when it
arises.