Implement sliding window algorithm for FMD params (#4473)
## Describe your changes

This adds a new algorithm for FMD parameters, which uses a sliding
window approximation to target a specified number of detections over a
given window. The idea is that if you set a window, of say, 1 day (in
blocks), then the algorithm will then approximate the number of clues
detected in the past day, and then use that to set the precision, so
that the lowest precision such that at least the target number of
detections will be generated, if that number of clues is seen over the
upcoming window.

We could use a window of 1 day and 1000 detections or so as reasonable
production parameters.

## Issue ticket number and link

Closes #1087.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This adds new possibilities to the chain parameters in a
backwards-compatible way, but which will cause a consensus divergence if
the new parameters are sent via governance etc. without an upgrade