feat(fee): add FromStr, Display impls for FeeTier
We recently enabled fees on Testnet 77 (#4306), and in the process
broke a few things, such as Galileo. While working on adding FeeTier
support to Galileo, Galileo's clap setup wanted Display and FromStr.
Added those to satisfy the build, and then circled back to pcli to reuse
the same impls, which allowed us to snip out some duplicative code.