ci: support point-releases to solo node
The solo node, run outside the cluster, was being fully reset on
point releases, due to the aggressive reset logic that was written
during the days of frequent chain resets. We're performing
point-releases more regularly now, so let's make sure to avoid resetting
the solo node in those cases. Still unsupported is chain-migration logic
specifically for the solo node, but that's fine: we're handling chain
upgrades out of band for now anyway.

Closes #4459.