ci: don't run long tests on docs-only PRs
I want to keep the docs up to date, and sometimes that requires fussy
management of version numbers, particularly as we invest in multi-branch
workflows to support point releases (e.g. #4545). If *only* the docs
changed, skip the long-running rust tests, because we don't need to
validate rust functionality on docs changes.