Pretty print transaction plan in threshold custody using view (#4572)
This converts the plan into a view, using dummy values, so that we can
reuse the existing formatting.

This only handles outputs and spends, but that should be enough for 90%
of use cases, the rest of which can look at the JSON.

Closes #3444.

## Describe your changes

## Issue ticket number and link

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Just changes how pcli prints stuff to stdout