fix: metrics dev-env for linux
Fixes a hardcoded path in the compose setup for local metrics. Updates
the docs, as well, to clarify that the first-run experience via docker
compose is only relevant for Linux hosts, due to use of host-networking.
Will work on a macos-compatible dev env later, but for now focusing on
making sure what's in the repo is clear and runnable for actual testnet
node operators, which means Linux boxes.

Closes #4565.