pcli: 💿 crate can be used as a library
currently, types and interfaces defined in the workspace's `pcli` crate
are only available within the `pcli` binary itself. this means that the
client functionality provided by this crate cannot be depended on as a
library, or otherwise used in unit tests.

this state of affairs is a significant part of why our tests for pcli
are contained in the `crates/bin/pcli/tests/network_integration.rs` test
suite, which rely on shelling out via a `std::process::Command` wrapper
and invoking the compiled `pcli` binary.

to facilitate future unit testing, this commit introduces a `lib.rs`
file to the `pcli` crate. the `App` structure is moved there as part of
this change, but is not changed in this commit.