Define staking events (#4449)
## Describe your changes

Resolves #4336, emitting events for validator definitions and state
transitions. Included events:

- validator state change
- validator voting power change
- validator bonding state change
- validator rate data change
- validator definition upload
- validator missed block
- slashing penalty applied
- delegate
- undelegate
- funding stream rewards

All of these are proto-based events, and emitted both at genesis, when
relevant, and when changed later.

Emitting events **at** genesis is difficult because the
ABCI interface does not accept events as a response to `InitChain`, and we don't want to work around this fact as it violates the intent of event emission (to indicate changes to a previous state). This means the genesis block will emit no events -- indexers will need to process the `genesis.json` manually to obtain initial state.

## Issue ticket number and link

Resolves #4336. Closes #3797. Closes
#[4512](https://github.com/penumbra-zone/penumbra/issues/4512).

## Checklist before requesting a review

- [X] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > This adds new events, but does not affect consensus.

---------

Co-authored-by: Chris Czub <chris@penumbralabs.xyz>