proto: add spend to `TransactionPlannerRequest` (#4615)
## Describe your changes

previously, there wasn't an need to explicitly specify a _spend_ as a
user intent until non-zero fees were introduced. In that context, this
proto-related modification is necessary in order to allow a user to send
an _entire_ balance for a specific asset.

## Issue ticket number and link

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason: