chore: update required version for buf tool
The `buf` cli made a breaking change in `v1.32` [sic, 0], so we must
update the script invocations to keep it working. Touched up the docs
where appropriate, and ran the recommended `buf config migrate` cmd [1]
to get the files in sync with the new format. After rerunning `just
proto` on the migrated files, I see no changes.

[0] https://github.com/bufbuild/buf/releases/tag/v1.32.0
[1] https://buf.build/docs/migration-guides/migrate-v2-config-files

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>