metrics: add os-level cpu stats to pd
Pulls in the `metrics-process` crate [0] to bolt on CPU usage and other OS-level info
to the metrics emitted by pd. Doing so will support a better first-run
experience for node operators who lack a comprehensive pre-existing setup for monitoring node health,
so they can get a sense of how much load pd is under.

Adds a cursory first-pass on a new dashboard, but will need to follow up
on that once the new metrics actually land on hosts that are receiving
load. Also included a reference to metrics addd in #4581, as a treat.

[0] https://docs.rs/metrics-process/2.0.0/metrics_process/index.html