pd: 🙏 propagate errors from ACME worker
when auto-https is enabled, we spawn a task in the background to handle
https certificate resolution, via `rustls_acme::AcmeState`.

if that task encounters errors, they should be propagated up to the
daemon, so that `pd` does not rapidly retry lookups, potentially hitting
rate-limits and causing service interruptions.

this changes the `pd` entrypoint, binding the [`JoinHandle`] to a
variable and polling upon that future in the `tokio::select` block that
represents the core steady-state event loop of the daemon.

we also update the acme_worker loop to log-and-bail on error, ensuring
that pd exits when the error hits the select loop in pd main.

Co-Authored-By: Conor Schaefer <conor@penumbralabs.xyz>